Create NTNXvision.ps1
All you need to have is Office Visio 2013 on your workstation (haven't tested it with 2010). I hope it will help IT consultant avoiding some boring tasks during the customer report 
 
This script report all nodes composing your cluster, the containers and the VMs hosted on each of your hosts into a new Visio Sheet.
 
It recognize the Block model type of your hardware, the upperlying hypervisor type and the OS in your VM.
 
The script is composed of 2 files:
NTNX_Shape.vssx, a Visio 2013 Gabarit which contains the shapes. This files need to be placed in the same folder that the PS Script.
NTNXvision.ps1, the Powershell script that you need to launch to orchestrate all this stuff Smiley Happy

 
You can customize the diagram by simply changing some value in the script, if you would like to apply some Office Theme or anything.

See here for the original source http://next.nutanix.com/t5/Total-Recode/Dynamically-generate-your-Visio-Diagram/m-p/4935#U4935