Create deploy-nested-NTNXCE
Powershell-based GUI - deploying the Nutanix CE as a nested hypervisor on ESXi 

This script is a Powershell-based GUI to collect host, network, and storage configuration for deploying the Nutanix CE as a nested hypervisor on ESXi. The tool can deploy 1-4 nodes and each node can have a distinct configuration.
 
Summary of customizable parameters:
- Number of Nodes to Deploy
- VM Name Base: Node number appended to end Ex. NutanixCE --> NutanixCE1, NutanixCE2, NutanixCE3, NutanixCE4
- Port Group Name: A new port group is created with promiscuous mode enabled. This is required for a nested hypervisor on ESXi due to the way that MAC address tables are maintained.
- VLAN ID for Port Group: Can be an access or trunk port
- Path to CE disk image file
- Number of vCPUs
- RAM Capacity in GB
- ESXi Host Name
- vSwtich
- Acropolis Home Datastore
- SSD Datastore
- SSD Size
- HDD Datastore
- HDD Size
Upon executing the script, the user is prompted to enter connection information for vCenter or an ESXi host.
When credentials are entered, the connection can be completed by clicking the "Connect" button.
The ESXi host, vSwitch, and Datastore values are collected and populated in the dropdown menus for Nutanix node configuration.
When a valid configuration is completed, the Deploy Nutanix Nodes button can be pressed to complete the deployment.
After nodes are deployed, the Open VM Console button can be selected to open a VMRC connection in the web browser. (The VMware Client Plugin must be explicitly approved for automatic operation.)